# api/openapi.yaml
openapi: 3.1.0
info:
  title: llm-proxy web API
  version: 0.1.0
  description: |
    The browser-facing API of llm-proxy: the user cabinet and the admin panel.
    This is NOT the proxied LLM API (`/v1/...`), which the embedded upstream serves
    and which authenticates with API tokens, not sessions.

    Conventions:
    - Authentication is a session cookie set by `POST /api/auth/login` or the OIDC
      callback. It is `HttpOnly`, so no endpoint ever returns a session id.
    - A restricted session (temporary password not yet changed) may call only
      `GET /api/me`, `POST /api/auth/password` and `POST /api/auth/logout`; every
      other authenticated endpoint answers 403 with code `password_change_required`.
    - An authenticated endpoint called without a valid session answers 401 with code
      `unauthenticated`. The only other 401 code is `invalid_credentials`, and it is
      returned only by `POST /api/auth/login` and `POST /api/auth/password`, where it
      rejects what was typed and leaves the session alone: a client must not treat it
      as a signed-out session.
    - Every error the web API writes, including those from its middleware (session,
      restriction, content type), is a JSON `Error` - never plain text.
    - Codes any endpoint may answer besides those it declares: `unsupported_media_type`
      (415, a mutating request without `Content-Type: application/json`),
...
      expects, with `field` when one field is at fault).
...
servers:
  - url: /
...
paths:
...
  /api/admin/providers/login/start:
    post:
      tags: [admin-providers]
      operationId: startProviderLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderLoginStartRequest" }
      responses:
        "201":
          description: >
            A pending login. The administrator opens `authURL` in their own browser,
            completes the vendor sign-in, and pastes the URL they land on into
            `complete` before `expiresAt`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderLoginSession" }
        "409":
          description: "`login_busy` — too many vendor logins are pending; finish or abandon one and retry."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "`unsupported_provider`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/providers/login/complete:
    post:
      tags: [admin-providers]
      operationId: completeProviderLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderLoginCompleteRequest" }
      responses:
        "201":
          description: The account was added and is routable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderAccount" }
        "410":
          description: "`login_expired` — the session timed out or was already used."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "`login_failed` — the vendor rejected the callback."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/providers/{accountId}:
    patch:
      tags: [admin-providers]
      operationId: updateProviderAccount
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [disabled]
              properties:
                disabled: { type: boolean }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderAccount" }
    delete:
      tags: [admin-providers]
      operationId: removeProviderAccount
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      responses:
        "204": { description: Removed from memory and from storage. }
  /api/admin/settings:
    get:
      tags: [admin-settings]
      operationId: getSettings
      responses:
        "200":
          description: The gateway settings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
    put:
      tags: [admin-settings]
      operationId: updateSettings
      summary: >
        Validate, and unless `dryRun`, apply. The response always carries the diff
        against the running configuration, so the client can show it before applying.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsUpdateRequest" }
      responses:
        "200":
          description: Validated, and applied unless `dryRun`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SettingsUpdateResult" }
        "422":
          description: >
            `invalid_settings` or `forbidden_setting` (a field the gateway owns, such as
            the listen address, the auth directory, the management secret or config API
            keys), with `field` naming it.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/prices:
    get:
      tags: [admin-settings]
      operationId: getPrices
      responses:
        "200":
          description: The price list used to estimate cost in metrics.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ModelPrice" }
    put:
      tags: [admin-settings]
      operationId: replacePrices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/ModelPrice" }
      responses:
        "200":
          description: Replaced.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ModelPrice" }
        "422":
          description: >
            `invalid_input` with `field` naming the entry and value, e.g. `[2].input` (a
            negative price) or `[3]` (a provider and model listed twice). Nothing is replaced.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: llmproxy_session
  parameters:
...
    AccountId:
      name: accountId
      in: path
      required: true
      description: A vendor account id. Not a UUID — the upstream credential identifier.
      schema: { type: string }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine code. The client translates this, not `message`.
          examples: [invalid_credentials]
        message: { type: string, description: English diagnostic for logs. Not for display. }
        field: { type: string, description: "The request field the error is about, when there is one." }
...
    ProviderAccount:
      type: object
      required: [id, provider, status, disabled, quota]
      properties:
        id: { type: string }
        provider: { type: string, description: Policy-facing provider name. }
        label: { type: string }
        email: { type: [string, "null"] }
        status: { type: string, examples: [active, error, disabled] }
        disabled: { type: boolean }
        lastError: { type: [string, "null"] }
        lastRefreshedAt: { type: [string, "null"], format: date-time }
        quota:
          type: array
          description: Direct vendor signals. No capacity estimate.
          items:
            type: object
            required: [window, usedRatio]
            properties:
              window: { type: string, examples: [5h, 7d] }
              usedRatio: { type: number, minimum: 0, maximum: 1 }
              resetAt: { type: [string, "null"], format: date-time }
              observedAt: { type: [string, "null"], format: date-time }
    ProviderLoginStartRequest:
      type: object
      required: [provider]
      properties:
        provider: { type: string, enum: [claude, chatgpt] }
    ProviderLoginSession:
      type: object
      required: [sessionId, authURL, expiresAt]
      properties:
        sessionId: { type: string }
        authURL: { type: string, description: Open in the administrator's own browser. }
        expiresAt: { type: string, format: date-time }
    ProviderLoginCompleteRequest:
      type: object
      required: [sessionId, callbackURL]
      properties:
        sessionId: { type: string }
        callbackURL: { type: string, description: The full URL the vendor sign-in ended on. }
    Settings:
      type: object
      required: [yaml, fields]
      properties:
        yaml:
          type: string
          description: >
            The editable part of the upstream configuration as YAML. Fields the gateway
            owns (listen address, auth directory, management, config API keys) are
            absent and cannot be set.
        fields:
          type: object
          description: The same values as typed form fields.
          properties:
            proxyURL: { type: string }
            requestRetry: { type: integer, minimum: 0 }
            maxRetryInterval: { type: integer, minimum: 0, description: Seconds. }
    SettingsUpdateRequest:
      type: object
      properties:
        yaml: { type: string, description: Replaces the whole editable document. Mutually exclusive with `fields`. }
        fields:
          type: object
          description: Patches individual typed fields. Mutually exclusive with `yaml`.
          properties:
            proxyURL: { type: string }
            requestRetry: { type: integer, minimum: 0 }
            maxRetryInterval: { type: integer, minimum: 0 }
        dryRun: { type: boolean, default: false }
    SettingsUpdateResult:
      type: object
      required: [applied, diff, settings]
      properties:
        applied: { type: boolean }
        diff: { type: string, description: Unified diff of the YAML against the running configuration. }
        settings: { $ref: "#/components/schemas/Settings" }
    ModelPrice:
      type: object
      required: [provider, model, input, output, cacheRead, cacheWrite]
      description: US dollars per million tokens. An estimate of work done, not a bill.
      properties:
        provider: { type: string }
        model: { type: string }
        input: { type: number, format: double, minimum: 0 }
        output: { type: number, format: double, minimum: 0 }
        cacheRead: { type: number, format: double, minimum: 0 }
        cacheWrite: { type: number, format: double, minimum: 0 }
security:
  - session: []